appservice/requirements.txt
fastapi[all]
gunicorn==20.1.0
uvicorn==0.15.0
pymongo
azure-appconfiguration
azure-identity